Rename mislabelled packages
