+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/core/actions/fake/Configuration.java
@@ -1,4 +1,4 @@
-package ca.on.oicr.gsi.shesmu.ratelimit.actions.fake;
+package ca.on.oicr.gsi.shesmu.core.actions.fake;
 
 public class Configuration {
   private String allow;
