+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/core/package-info.java
@@ -4,4 +4,4 @@
  * <p>Shesmu is designed around a plug-in architecture. This package contains default functionality
  * that is included by default. Effectively, these are “built-in plug-ins”.
  */
-package ca.on.oicr.gsi.shesmu.ratelimit;
+package ca.on.oicr.gsi.shesmu.core;
