+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/core/signers/JsonSigner.java
@@ -1,4 +1,4 @@
-package ca.on.oicr.gsi.shesmu.ratelimit.signers;
+package ca.on.oicr.gsi.shesmu.core.signers;
 
 import ca.on.oicr.gsi.shesmu.plugin.json.PackJsonObject;
 import ca.on.oicr.gsi.shesmu.plugin.signature.DynamicSigner;
