+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/core/signers/SHA1DigestSigner.java
@@ -1,4 +1,4 @@
-package ca.on.oicr.gsi.shesmu.ratelimit.signers;
+package ca.on.oicr.gsi.shesmu.core.signers;
 
 import ca.on.oicr.gsi.shesmu.plugin.signature.DynamicSigner;
 import ca.on.oicr.gsi.shesmu.plugin.types.Field;
